By @h0daka

+++ VegetaRobot/modules/sql/approve_sql.py
@@ -1,14 +1,14 @@
 import threading
 
-from sqlalchemy import Column, String, UnicodeText, Integer, func, distinct
-
+from sqlalchemy import Column, String, Integer
+from sqlalchemy.sql.sqltypes import BigInteger
 from VegetaRobot.modules.sql import BASE, SESSION
 
 
 class Approvals(BASE):
     __tablename__ = "approval"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
 
     def __init__(self, chat_id, user_id):
         self.chat_id = str(chat_id)  # ensure string
